Fix the global startid.
Previously, we would process the whole database every time (rejecting all the
previously archived conversations row-by-row. Now we can skip most of the
database during our select query if we've run through it previously.

This still has to manually reject rows if there was a conversation ongoing in
the last 30 minutes on the prior run, but we're talking about rejecting a
handful of rows instead of 100,000 of them.